agreg.fit {survival}	R Documentation

Cox model fitting functions

Description

These are the the functions called by coxph that do the actual computation. In certain situations, e.g. a simulation, it may be advantageous to call these directly rather than the usual coxph call using a model formula.

Usage


agreg.fit(x, y, strata, offset, init, control, weights, method,
rownames, resid=TRUE, concordance=TRUE)
coxph.fit(x, y, strata, offset, init, control, weights, method,
rownames, resid=TRUE, concordance=TRUE)

Arguments

x
Matix of predictors. This should not include an intercept.

y
a Surv object containing either 2 columns (coxph.fit) or 3 columns (agreg.fit).

strata
a vector containing the stratification, or NULL

offset
optional offset vector

init
initial values for the coefficients

control
the result of a call to coxph.control

weights
optional vector of weights

method
method for hanling ties, one of "breslow" or "efron"

rownames
this is only needed for a NULL model, in which case it contains the rownames (if any) of the original data.

resid
compute and return residuals.

concordance
compute and return the concordance.

Details

This routine does no checking that arguments are the proper length or type. Only use it if you know what you are doing!

The resid and concordance arguments will save some compute time for calling routines that only need the likelihood, the generation of a permutation distribution for instance.

Value

a list containing results of the fit

Author(s)

Terry Therneau

See Also

coxph

[Package survival version 2.42-6 Index]