as.date {survival}	R Documentation

Coerce Data to Dates

Description

Converts any of the following character forms to a Julian date: 8/31/56, 8-31-1956, 31 8 56, 083156, 31Aug56, or August 31 1956.

Usage

as.date(x, order = "mdy", ...)

Arguments

x 	 input data vector.

order 	 if x is character, defines the order in which the terms are assumed to appear in a xx/xx/xx date. The default is month/day/year; any permutation of mdy is legal.

... 	 if x is character, then any other arguments from mdy.date() can be used as well.

Details

If x is numeric, then floor(x) is returned, e.g., as.date(35) is the same as as.date(35.2) and gives Feb 2, 1960. If x is character, the program attempts to parse it.

Value

For each date, the number of days between it and January 1, 1960. The date will be missing if the string is not interpretable.

See Also

mdy.date, date.mmddyy, date.ddmmmyy

Examples

as.date(c("1jan1960", "2jan1960", "31mar1960", "30jul1960"))

[Package survival version 2.34-1 Index]