concordancefit {survival}	R Documentation

Compute the concordance

Description

This is the working routine behind the concordance function. It is not meant to be called by users, but is available for other packages to use. Input arguments, for instance, are assumed to all be the correct length and type, and missing values are not allowed: the calling routine is responsible for these things.

Usage

concordancefit(y, x, strata, weights, ymin = NULL, ymax = NULL,
 timewt = c("n", "S", "S/G", "n/G", "n/G2", "I"), cluster, influence =0,
 ranks = FALSE, reverse = FALSE, timefix = TRUE, keepstrata=10)

Arguments

y
the response. It can be numeric, factor, or a Surv object

x
the predictor, a numeric vector

strata
optional numeric vector that stratifies the data

weights
options vector of case weights

ymin, ymax
restrict the comparison to response values in this range

timewt
the time weighting to be used

cluster, influence,ranks, reverse, timefix
see the help for the concordance function

keepstrata
either TRUE, FALSE, or an integer value. Computations are always done within stratum, then added. If the total number of strata greater than keepstrata, or keepstrata=FALSE, those subtotals are not kept in the output.

Value

a list containing the results

Author(s)

Terry Therneau

See Also

concordance

[Package survival version 3.1-12 Index]