coxph.wtest {survival}	R Documentation

Compute a quadratic form

Description

This function is used internally by several survival routines. It computes a simple quadratic form, while properly dealing with missings.

Usage


coxph.wtest(var, b, toler.chol = 1e-09)

Arguments

var
variance matrix

b
vector

toler.chol
tolerance for the internal cholesky decomposition

Details

Compute b' V-inverse b. Equivalent to sum(b * solve(V,b)), except for the case of redundant covariates in the original model, which lead to NA values in V and b.

Value

a real number

Author(s)

Terry Therneau

[Package survival version 2.42-6 Index]