date.mmddyy {survival}	R Documentation

Format a Julian date

Description

Given a vector of Julian dates, this returns them in the form ``10/11/89'', ``28/7/54'', etc.

Usage

date.mmddyy(sdate, sep = "/")

Arguments

sdate 	 A vector of Julian dates, e.g., as returned by mdy.date().

sep 	 Character string used to separate the month, day, and year portions of the returned string.

Value

A vector of character strings containing the formatted dates.

See Also

date.mdy, mdy.date, date.ddmmmyy

Examples

date.mmddyy(as.date(10))

[Package survival version 2.34-1 Index]