date.object {survival}	R Documentation

Date Objects

Description

Objects of class "date".

Usage

is.date(x)

Arguments

x 	 any R object.

Details

Dates are stored as the number of days since 1/1/1960, and are kept in integer format. (This is the same baseline value as is used by SAS). The numerical methods for dates treat date - date as a numeric, and date +- numeric as a date.

is.date returns TRUE if x has class "date", and FALSE otherwise. Its behavior is unaffected by any attributes of x; for example, x could be a date array (in contrast to the behavior of is.vector).

as.date returns x if x is a simple object of class "date", and otherwise a date vector of the same length as x and with data resulting from coercing the elements of x to class "date". See the manual page for as.date() for details.

Logical operations as well as the numeric functions exp(), log(), and so on are invalid.

Other methods exist for missing value, as.character(), printing, and summarizing.

See Also

date.mdy, mdy.date, date.ddmmmyy, as.date.

[Package survival version 2.34-1 Index]