is.ratetable {survival}	R Documentation

Verify that an object is of class ratetable.

Description

The function verifies not only the class attribute, but the structure of the object.

Usage


is.ratetable(x, verbose=FALSE)

Arguments

x
the object to be verified.

verbose
if TRUE and the object is not a ratetable, then return a character string describing the way(s) in which x fails to be a proper ratetable object.

Details

Rate tables are used by the pyears and survexp functions, and normally contain death rates for some population, categorized by age, sex, or other variables. They have a fairly rigid structure, and the verbose option can help in creating a new rate table.

Value

returns TRUE if x is a ratetable, and FALSE or a description if it is not.

See Also

pyears, survexp.

Examples


is.ratetable(survexp.us)  # True
is.ratetable(cancer)      # False

[Package survival version 2.42-6 Index]