kidney {survival}	R Documentation

Kidney data from survival5

Description

Data on the recurrence times to infection, at the point of insertion of the catheter, for kidney patients using portable dialysis equipment. Catheters may be removed for reasons other than infection, in which case the observation is censored. Each patient has exactly 2 observations.

Format

patient: 	 id

time: 	 time

status: 	 event status

age: 	 in years

sex: 	 1=male, 2=female

disease: 	 disease type (0=GN, 1=AN, 2=PKD, 3=Other)

frail: 	 frailty estimate from original paper

Note

The original analysis had incorrect handling of ties and so is not exactly reproduced by survival.

Source

CA McGilchrist, CW Aisbett (1991), Regression with frailty in survival analysis. Biometrics 47, 461–66.

Examples

kfit <- coxph(Surv(time, status)~ age + sex + disease + frailty(id), kidney)
kfit0 <- coxph(Surv(time, status)~ age + sex + disease, kidney)
kfitm1 <- coxph(Surv(time,status) ~ age + sex + disease + 
                frailty(id, dist='gauss'), kidney)

[Package survival version 2.34-1 Index]