levels.Surv {survival}	R Documentation

Return the states of a multi-state Surv object

Description

For a multi-state Surv object, this will return the names of the states.

Usage


## S3 method for class 'Surv'
levels(x)

Arguments

x
a Surv object

Value

for a multi-state Surv object, the vector of state names (excluding censoring); or NULL for an ordinary Surv object

Examples


y1 <- Surv(c(1,5, 9, 17,21, 30),
           factor(c(0, 1, 2,1,0,2), 0:2, c("censored", "progression", "death")))
levels(y1)

y2 <- Surv(1:6, rep(0:1, 3))
y2
levels(y2)

[Package survival version 2.42-6 Index]