model.frame.coxph {survival}	R Documentation

Model.frame method for coxph objects

Description

Recreate the model frame of a coxph fit.

Usage


## S3 method for class 'coxph'
model.frame(formula, ...)

Arguments

formula
the result of a coxph fit

...
other arguments to model.frame

Details

For details, see the manual page for the generic function. This function would rarely be called by a user, it is mostly used inside functions like residual that need to recreate the data set from a model in order to do further calculations.

Value

the model frame used in the original fit, or a parallel one for new data.

Author(s)

Terry Therneau

See Also

model.frame

[Package survival version 2.42-6 Index]