ratetableDate {survival}	R Documentation

Convert date objects to ratetable form

Description

This method converts dates from various forms into the internal form used in ratetable objects.

Usage


ratetableDate(x)

Arguments

x
a date. The function currently has methods for Date, date, POSIXt, timeDate, and chron objects.

Details

This function is useful for those who create new ratetables, but is normally invisible to users. It is used internally by the survexp and pyears functions to map the various date formats; if a new method is added then those routines will automatically be adapted to the new date type.

Value

a numeric vector, the number of days since 1/1/1960.

Author(s)

Terry Therneau

See Also

pyears, survexp

[Package survival version 2.42-6 Index]