summary.survexp {survival}	R Documentation

Summary function for a survexp object

Description

Returns a list containing the values of the survival at specified times.

Usage


## S3 method for class 'survexp'
summary(object, times, scale = 1, ...)

Arguments

object
the result of a call to the survexp function

times
vector of times; the returned matrix will contain 1 row for each time. Missing values are not allowed.

scale
numeric value to rescale the survival time, e.g., if the input data to survfit were in days, scale = 365.25 would scale the output to years.

...
For future methods

Details

A primary use of this function is to retrieve survival at fixed time points, which will be properly interpolated by the function.

Value

a list with the following components:

surv
the estimate of survival at time t.

time
the timepoints on the curve.

n.risk
In expected survival each subject from the data set is matched to a hypothetical person from the parent population, matched on the characteristics of the parent population. The number at risk is the number of those hypothetical subject who are still part of the calculation.

Author(s)

Terry Therneau

See Also

survexp

[Package survival version 2.42-6 Index]