summary.survfit {survival}	R Documentation

Summary of a Survival Curve

Description

Returns a list containing the survival curve, confidence limits for the curve, and other information.

Usage


  ## S3 method for class 'survfit'
summary(object, times, censored=FALSE, scale=1,
  extend=FALSE, rmean=getOption('survfit.rmean'), ...)
  

Arguments

object
the result of a call to the survfit function.

times
vector of times; the returned matrix will contain 1 row for each time. The vector will be sorted into increasing order; missing values are not allowed. If censored=T, the default times vector contains all the unique times in fit, otherwise the default times vector uses only the event (death) times.

censored
logical value: should the censoring times be included in the output? This is ignored if the times argument is present.

scale
numeric value to rescale the survival time, e.g., if the input data to survfit were in days, scale = 365.25 would scale the output to years.

extend
logical value: if TRUE, prints information for all specified times, even if there are no subjects left at the end of the specified times. This is only valid if the times argument is present.

rmean
Show restricted mean: see print.survfit for details

...
for future methods

Value

a list with the following components:

surv
the estimate of survival at time t+0.

time
the timepoints on the curve.

n.risk
the number of subjects at risk at time t-0 (but see the comments on weights in the survfit help file).

n.event
if the times argument is missing, then this column is the number of events that occurred at time t. Otherwise, it is the cumulative number of events that have occurred since the last time listed until time t+0.

n.entered
This is present only for counting process survival data. If the times argument is missing, this column is the number of subjects that entered at time t. Otherwise, it is the cumulative number of subjects that have entered since the last time listed until time t.

n.exit.censored
if the times argument is missing, this column is the number of subjects that left without an event at time t. Otherwise, it is the cumulative number of subjects that have left without an event since the last time listed until time t+0. This is only present for counting process survival data.

std.err
the standard error of the survival value.

conf.int
level of confidence for the confidence intervals of survival.

lower
lower confidence limits for the curve.

upper
upper confidence limits for the curve.

strata
indicates stratification of curve estimation. If strata is not NULL, there are multiple curves in the result and the surv, time, n.risk, etc. vectors will contain multiple curves, pasted end to end. The levels of strata (a factor) are the labels for the curves.

call
the statement used to create the fit object.

na.action
same as for fit, if present.

table
table of information that is returned from print.survfit function.

type
type of data censoring. Passed through from the fit object.

See Also

survfit, print.summary.survfit

Examples


summary( survfit( Surv(futime, fustat)~1, data=ovarian))
summary( survfit( Surv(futime, fustat)~rx, data=ovarian))

[Package survival version 2.42-6 Index]