survexp.fit {survival}	R Documentation

Compute Expected Survival

Description

Compute expected survival times.

Usage


survexp.fit(group, x, y, times, death, ratetable)

Arguments

group
if there are multiple survival curves this identifies the group, otherwise it is a constant. Must be an integer.

x
A matrix whose columns match the dimensions of the ratetable, in the correct order.

y
the follow up time for each subject.

times
the vector of times at which a result will be computed.

death
a logical value, if TRUE the conditional survival is computed, if FALSE the cohort survival is computed. See survexp for more details.

ratetable
a rate table, such as survexp.uswhite.

Details

For conditional survival y must be the time of last follow-up or death for each subject. For cohort survival it must be the potential censoring time for each subject, ignoring death.

For an exact estimate times should be a superset of y, so that each subject at risk is at risk for the entire sub-interval of time. For a large data set, however, this can use an inordinate amount of storage and/or compute time. If the times spacing is more coarse than this, an actuarial approximation is used which should, however, be extremely accurate as long as all of the returned values are > .99.

For a subgroup of size 1 and times > y, the conditional method reduces to exp(-h) where h is the expected cumulative hazard for the subject over his/her observation time. This is used to compute individual expected survival.

Value

A list containing the number of subjects and the expected survival(s) at each time point. If there are multiple groups, these will be matrices with one column per group.

Warning

Most users will call the higher level routine survexp. Consequently, this function has very few error checks on its input arguments.

See Also

survexp, survexp.us.

[Package survival version 2.42-6 Index]