survfit.object {survival}	R Documentation

Survival Curve Object

Description

This class of objects is returned by the survfit class of functions to represent a fitted survival curve. For a multi-state model the object has class c('survfitms', 'survfit').

Objects of this class have methods for the functions print, summary, plot, points and lines. The print.survfit method does more computation than is typical for a print method and is documented on a separate page.

Arguments

n
total number of subjects in each curve.

time
the time points at which the curve has a step.

n.risk
the number of subjects at risk at t.

n.event
the number of events that occur at time t.

n.enter
for counting process data only, the number of subjects that enter at time t.

n.censor
for counting process data only, the number of subjects who exit the risk set, without an event, at time t. (For right censored data, this number can be computed from the successive values of the number at risk).

surv
the estimate of survival at time t+0. This may be a vector or a matrix. The latter occurs when a set of survival curves is created from a single Cox model, in which case there is one column for each covariate set.

prev, p0
a multi-state survival will have the prev component instead of surv. It will be a matrix containing the estimated probability of each state at each time, one column per state. The p0 matrix contains the initial distribution of states. (On further reflection pstate= "probability in state" would have been a much better label than "prevalence", but by that point too many other packages were dependent on the form of the result.)

std.err
for a survival curve this contains standard error of the cumulative hazard or -log(survival), for a multi-state curve it contains the standard error of prev. This difference is a reflection of the fact that each is the natural calculation for that case.

cumhaz hazard
optional. For a multi-state curve this is an k by k array for each time point, where k is the number of states.

upper
upper confidence limit for the survival curve or probability

lower
lower confidence limit for the survival curve or probability

strata
if there are multiple curves, this component gives the number of elements of the time etc. vectors corresponding to the first curve, the second curve, and so on. The names of the elements are labels for the curves.

start.time
the value specified for the start.time argument, if it was used in the call.

n.all
for counting process data, and any time that the start.time argument was used, this contains the total number of observations that were available. Not all may have been used in creating the curve, in which case this value will be larger than n above.

conf.type
the approximation used to compute the confidence limits.

conf.int
the level of the confidence limits, e.g. 90 or 95%.

transitions
for multi-state data, the total number of transitions of each type.

na.action
the returned value from the na.action function, if any. It will be used in the printout of the curve, e.g., the number of observations deleted due to missing values.

call
an image of the call that produced the object.

type
type of survival censoring.

influence
optional, for survfitms objects only. A list with one element per stratum, each element of the list is an array indexed by subject, time, state. The time dimension will have one more element than the prev matrix, the first row is the subject's influence on the initial prevalence (just before the first time point). If there is only one curve a list is not needed.

Structure

The following components must be included in a legitimate survfit or survfitms object.

Subscripts

Survfit objects that contain multiple survival curves can be subscripted. This is often used to plot a subset of the curves. If the surv or prev component is a matrix then the survfit object will be treated as a matrix, otherwise only a single subscript is used.

See Also

plot.survfit, summary.survfit, print.survfit, survfit.

[Package survival version 2.42-6 Index]