survival-deprecated {survival}	R Documentation

Deprecated functions in package survival

Description

These functions are temporarily retained for compatability with older programs, and may transition to defunct status.

Usage

survConcordance(formula, data, weights, subset, na.action) # use concordance
survConcordance.fit(y, x, strata, weight)    # use concordancefit

Arguments

formula
a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.

data
a data frame

weights,subset,na.action
as for coxph

x, y, strata, weight
predictor, response, strata, and weight vectors for the direct call

See Also

Deprecated

[Package survival version 3.3-1 Index]