survreg.old {survival}	R Documentation

Old (survival4) Regression for a parametric survival model

Description

This routine is a backwards-compatible interface to the improved survreg function, which is better.

Usage

survreg.old(formula, data=sys.frame(sys.parent()), ..., link=c("log",
"identity"),dist=c("extreme", "logistic", "gaussian",
"exponential","rayleigh","weibull"), fixed=list())

Arguments

formula 	 a formula expression as for other regression models. See the documentation for lm and formula for details.

data 	 optional data frame in which to interpret the variables occuring in the formula.

... 	 other arguments to survreg

link 	 transformation to be used on the y variable.

dist 	 assumed distribution for the transformed y variable.

fixed 	 a list of fixed parameters, most often just the scale.

Value

an object of class survreg is returned, which inherits from class glm.

Examples

survreg.old(Surv(futime, fustat) ~ ecog.ps + rx, ovarian, dist='extreme',
                link='log', fixed=list(scale=1))   #Fit an exponential

[Package survival version 2.34-1 Index]