tobin {survival}	R Documentation

Tobin's Tobit data

Description

Economists fit a parametric censored data model called the `tobit'. These data are from Tobin's original paper.

Usage

data(tobin)

Format

A data frame with 20 observations on the following 3 variables.

durable

Durable goods purchase

age

Age in years

quant

Liquidity ratio (x 1000)

Source

J. Tobin, Estimation of relationships for limited dependent variables, Econometrica, v26, 24-36, 1958.

Examples

tfit <- survreg(Surv(durable, durable>0, type='left') ~age + quant,
                data=tobin, dist='gaussian')

predict(tfit,type="response")

[Package survival version 2.17 Index]