Analysis of Deviance for a Cox model.

Compute an analysis of deviance table for one or more Cox model fits, based on the log partial likelihood.

Usage

# S3 method for coxph
anova(object, ...,  test = 'Chisq')

Arguments

object

An object of class coxph

...

Further coxph objects

test

a character string. The appropriate test is a chisquare, all other choices result in no test being done.

Details

Specifying a single object gives a sequential analysis of deviance table for that fit. That is, the reductions in the model Cox log-partial-likelihood as each term of the formula is added in turn are given in as the rows of a table, plus the log-likelihoods themselves. A robust variance estimate is normally used in situations where the model may be mis-specified, e.g., multiple events per subject. In this case a comparison of likelihood values does not make sense (differences no longer have a chi-square distribution), and anova will refuse to print results.

If more than one object is specified, the table has a row for the degrees of freedom and loglikelihood for each model. For all but the first model, the change in degrees of freedom and loglik is also given. (This only make statistical sense if the models are nested.) It is conventional to list the models from smallest to largest, but this is up to the user.

The table will optionally contain test statistics (and P values) comparing the reduction in loglik for each row.

Value

An object of class "anova" inheriting from class "data.frame".

Warning

The comparison between two or more models by anova will only be valid if they are fitted to the same dataset. This may be a problem if there are missing values.

See also

Examples

fit <- coxph(Surv(futime, fustat) ~ resid.ds *rx + ecog.ps, data = ovarian) 
anova(fit)
#> Analysis of Deviance Table
#>  Cox model: response is Surv(futime, fustat)
#> Terms added sequentially (first to last)
#> 
#>              loglik  Chisq Df Pr(>|Chi|)  
#> NULL        -34.985                       
#> resid.ds    -33.105 3.7594  1    0.05251 .
#> rx          -32.269 1.6733  1    0.19582  
#> ecog.ps     -31.970 0.5980  1    0.43934  
#> resid.ds:rx -30.946 2.0469  1    0.15251  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
fit2 <- coxph(Surv(futime, fustat) ~ resid.ds +rx + ecog.ps, data=ovarian)
anova(fit2,fit)
#> Analysis of Deviance Table
#>  Cox model: response is  Surv(futime, fustat)
#>  Model 1: ~ resid.ds + rx + ecog.ps
#>  Model 2: ~ resid.ds * rx + ecog.ps
#>    loglik  Chisq Df Pr(>|Chi|)
#> 1 -31.970                     
#> 2 -30.946 2.0469  1     0.1525