basehaz {survival}	R Documentation

Alias for the survfit function

Description

Compute the predicted survival curve for a Cox model.

Usage

basehaz(fit, centered=TRUE)

Arguments

fit
a coxph fit

centered
if TRUE return data from a predicted survival curve at the mean values of the covariates fit$mean, if FALSE return a prediction for all covariates equal to zero.

Details

This function is simply an alias for survfit, which does the actual work and has a richer set of options. The alias exists only because some users look for predicted survival estimates under this name.

The function returns a data frame containing the time, cumhaz and optionally the strata (if the fitted Cox model used a strata statement), which are copied the survfit result. Results for all covariates =0 are a standard form found in textbooks, however, due to possible overflow in the exp() function this is, in practice, a very bad idea in practice.

Value

a data frame with variable names of hazard, time and optionally strata. The first is actually the cumulative hazard.

See Also

survfit.coxph

[Package survival version 3.4-0 Index]