bladder {survival}	R Documentation

Bladder Cancer Recurrences

Description

Data on recurrences of bladder cancer, used by many people to demonstrate methodology for recurrent event modelling.

Usage

data(bladder)

Format

bladder

id: 	 Patient id

rx: 	 Treatment 1=placebo 2=thiotepa

number: 	 Initial number of tumours (8=8 or more)

size: 	 size (cm) of largest initial tumour

stop: 	 recurrence or censoring time

enum: 	 which recurrence (up to 4)

bladder2

id: 	 Patient id

rx: 	 Treatment 1=placebo 2=thiotepa

number: 	 Initial number of tumours (8=8 or more)

size: 	 size (cm) of largest initial tumour

start: 	 start of interval (0 or previous recurrence time)

stop: 	 recurrence or censoring time

enum: 	 which recurrence (up to 4)

Source

Wei, Lin, Weisfeld, JASA 1989

[Package survival version 2.17 Index]