bladder {survival}	R Documentation

Bladder Cancer Recurrences

Description

Data on recurrences of bladder cancer, used by many people to demonstrate methodology for recurrent event modelling.

Usage

bladder
bladder2

Format

bladder

id: 	 Patient id

rx: 	 Treatment 1=placebo 2=thiotepa

number: 	 Initial number of tumours (8=8 or more)

size: 	 size (cm) of largest initial tumour

stop: 	 recurrence or censoring time

enum: 	 which recurrence (up to 4)

bladder2

id: 	 Patient id

rx: 	 Treatment 1=placebo 2=thiotepa

number: 	 Initial number of tumours (8=8 or more)

size: 	 size (cm) of largest initial tumour

start: 	 start of interval (0 or previous recurrence time)

stop: 	 recurrence or censoring time

enum: 	 which recurrence (up to 4)

Source

LJ Wei, DY Lin, L Weissfeld (1989), Regression analysis of multivariate incomplete failure time data by modeling marginal distributions. Journal of the American Statistical Association, 84.

[Package survival version 2.34-1 Index]