Fits proportional hazards regression model to case-cohort data

Returns estimates and standard errors from relative risk regression fit to data from case-cohort studies. A choice is available among the Prentice, Self-Prentice and Lin-Ying methods for unstratified data. For stratified data the choice is between Borgan I, a generalization of the Self-Prentice estimator for unstratified case-cohort data, and Borgan II, a generalization of the Lin-Ying estimator.

Usage

cch(formula, data, subcoh, id, stratum=NULL, cohort.size,
    method =c("Prentice","SelfPrentice","LinYing","I.Borgan","II.Borgan"),
    robust=FALSE)

Arguments

formula

A formula object that must have a Surv object as the response. The Surv object must be of type "right", or of type "counting".

subcoh

Vector of indicators for subjects sampled as part of the sub-cohort. Code 1 or TRUE for members of the sub-cohort, 0 or FALSE for others. If data is a data frame then subcoh may be a one-sided formula.

id

Vector of unique identifiers, or formula specifying such a vector.

stratum

A vector of stratum indicators or a formula specifying such a vector

cohort.size

Vector with size of each stratum original cohort from which subcohort was sampled

data

An optional data frame in which to interpret the variables occurring in the formula.

method

Three procedures are available. The default method is "Prentice", with options for "SelfPrentice" or "LinYing".

robust

For "LinYing" only, if robust=TRUE, use design-based standard errors even for phase I

Value

An object of class "cch" incorporating a list of estimated regression coefficients and two estimates of their asymptotic variance-covariance matrix.

coef

regression coefficients.

naive.var

Self-Prentice model based variance-covariance matrix.

var

Lin-Ying empirical variance-covariance matrix.

Details

Implements methods for case-cohort data analysis described by Therneau and Li (1999). The three methods differ in the choice of "risk sets" used to compare the covariate values of the failure with those of others at risk at the time of failure. "Prentice" uses the sub-cohort members "at risk" plus the failure if that occurs outside the sub-cohort and is score unbiased. "SelfPren" (Self-Prentice) uses just the sub-cohort members "at risk". These two have the same asymptotic variance-covariance matrix. "LinYing" (Lin-Ying) uses the all members of the sub-cohort and all failures outside the sub-cohort who are "at risk". The methods also differ in the weights given to different score contributions.

The data argument must not have missing values for any variables in the model. There must not be any censored observations outside the subcohort.

Author

Norman Breslow, modified by Thomas Lumley

References

Prentice, RL (1986). A case-cohort design for epidemiologic cohort studies and disease prevention trials. Biometrika 73: 1--11.

Self, S and Prentice, RL (1988). Asymptotic distribution theory and efficiency results for case-cohort studies. Annals of Statistics 16: 64--81.

Lin, DY and Ying, Z (1993). Cox regression with incomplete covariate measurements. Journal of the American Statistical Association 88: 1341--1349.

Barlow, WE (1994). Robust variance estimation for the case-cohort design. Biometrics 50: 1064--1072

Therneau, TM and Li, H (1999). Computing the Cox model for case-cohort designs. Lifetime Data Analysis 5: 99--112.

Borgan, \(O\), Langholz, B, Samuelsen, SO, Goldstein, L and Pogoda, J (2000) Exposure stratified case-cohort designs. Lifetime Data Analysis 6, 39-58.

See also

Examples

## The complete Wilms Tumor Data 
## (Breslow and Chatterjee, Applied Statistics, 1999)
## subcohort selected by simple random sampling.
##

subcoh <- nwtco$in.subcohort
selccoh <- with(nwtco, rel==1|subcoh==1)
ccoh.data <- nwtco[selccoh,]
ccoh.data$subcohort <- subcoh[selccoh]
## central-lab histology 
ccoh.data$histol <- factor(ccoh.data$histol,labels=c("FH","UH"))
## tumour stage
ccoh.data$stage <- factor(ccoh.data$stage,labels=c("I","II","III","IV"))
ccoh.data$age <- ccoh.data$age/12 # Age in years

##
## Standard case-cohort analysis: simple random subcohort 
##

fit.ccP <- cch(Surv(edrel, rel) ~ stage + histol + age, data =ccoh.data,
   subcoh = ~subcohort, id=~seqno, cohort.size=4028)

fit.ccP
#> Case-cohort analysis,x$method, Prentice 
#>  with subcohort of 668 from cohort of 4028 
#> 
#> Call: cch(formula = Surv(edrel, rel) ~ stage + histol + age, data = ccoh.data, 
#>     subcoh = ~subcohort, id = ~seqno, cohort.size = 4028)
#> 
#> Coefficients:
#>               Value         SE        Z            p
#> stageII  0.73457084 0.16849620 4.359569 1.303187e-05
#> stageIII 0.59708356 0.17345094 3.442377 5.766257e-04
#> stageIV  1.38413197 0.20481982 6.757803 1.400990e-11
#> histolUH 1.49806307 0.15970515 9.380180 0.000000e+00
#> age      0.04326787 0.02373086 1.823274 6.826184e-02

fit.ccSP <- cch(Surv(edrel, rel) ~ stage + histol + age, data =ccoh.data,
   subcoh = ~subcohort, id=~seqno, cohort.size=4028, method="SelfPren")

summary(fit.ccSP)
#> Case-cohort analysis,x$method, SelfPrentice 
#>  with subcohort of 668 from cohort of 4028 
#> 
#> Call: cch(formula = Surv(edrel, rel) ~ stage + histol + age, data = ccoh.data, 
#>     subcoh = ~subcohort, id = ~seqno, cohort.size = 4028, method = "SelfPren")
#> 
#> Coefficients:
#>           Coef    HR  (95%   CI)     p
#> stageII  0.736 2.088 1.501 2.905 0.000
#> stageIII 0.597 1.818 1.294 2.553 0.001
#> stageIV  1.392 4.021 2.692 6.008 0.000
#> histolUH 1.506 4.507 3.295 6.163 0.000
#> age      0.043 1.044 0.997 1.094 0.069

##
## (post-)stratified on instit
##
stratsizes<-table(nwtco$instit)
fit.BI<- cch(Surv(edrel, rel) ~ stage + histol + age, data =ccoh.data,
   subcoh = ~subcohort, id=~seqno, stratum=~instit, cohort.size=stratsizes,
   method="I.Borgan")

summary(fit.BI)
#> Exposure-stratified case-cohort analysis, I.Borgan method.
#>              1   2
#> subcohort  952 202
#> cohort    3622 406
#> Call: cch(formula = Surv(edrel, rel) ~ stage + histol + age, data = ccoh.data, 
#>     subcoh = ~subcohort, id = ~seqno, stratum = ~instit, cohort.size = stratsizes, 
#>     method = "I.Borgan")
#> 
#> Coefficients:
#>           Coef    HR  (95%   CI)     p
#> stageII  0.737 2.090 1.501 2.909 0.000
#> stageIII 0.602 1.825 1.301 2.561 0.000
#> stageIV  1.395 4.036 2.702 6.029 0.000
#> histolUH 1.522 4.580 3.450 6.080 0.000
#> age      0.043 1.044 0.996 1.093 0.072