clogit {survival}	R Documentation

Conditional logistic regression

Description

Estimates a logistic regression model by maximising the conditional likelihood. Uses a model formula of the form case.status~exposure+strata(matched.set). The default is to use the exact conditional likelihood, a commonly used approximate conditional likelihood is provided for compatibility with older software.

Usage

clogit(formula, data, method=c("exact", "approximate"), na.action=getOption("na.action"), subset=NULL,control=coxph.control())

Arguments

formula 	 Model formula

data 	 data frame

method 	 correct or approximate conditonal likelihood?

na.action 	 missing value handling

subset 	 subset of records to use

control 	 control values

Value

An object of class "clogit", which is a wrapper for a "coxph" object.

Author(s)

Thomas Lumley

See Also

strata,coxph,glm

Examples

## Not run: clogit(case~spontaneous+induced+strata(stratum),data=infert)

[Package survival version 2.35-4 Index]