cluster {survival}	R Documentation

Identify clusters.

Description

This is a special function used in the context of survival models. It identifies correlated groups of observations, and is used on the right hand side of a formula. Using cluster() in a formula implies that robust sandwich variance estimators are desired.

Usage

cluster(x)

Arguments

x 	 A character, factor, or numeric variable.

Details

The function's only action is semantic, to mark a variable as the cluster indicator. The resulting variance is what is known as the ``working independence'' variance in a GEE model.

Value

x

See Also

coxph, survreg

Examples

marginal.model <- coxph(Surv(time, status) ~ rx + cluster(litter), rats)
frailty.model  <- coxph(Surv(time, status) ~ rx + frailty(litter), rats)

[Package survival version 2.35-4 Index]