colon {survival}	R Documentation

Chemotherapy for Stage B/C colon cancer

Description

These are data from one of the first successful trials of adjuvant chemotherapy for colon cancer. Levamisole is a low-toxicity compound previously used to treat worm infestations in animals; 5-FU is a moderately toxic (as these things go) chemotherapy agent. There are two records per person, one for recurrence and one for death

Usage

colon

Format

id: 	 id

study: 	 1 for all patients

rx: 	 Treatment - Obs(ervation), Lev(amisole), Lev(amisole)+5-FU

sex: 	 1=male

age: 	 in years

obstruct: 	 obstruction of colon by tumour

perfor: 	 perforation of colon

adhere: 	 adherence to nearby organs

nodes: 	 number of lymph nodes with detectable cancer

time: 	 days until event or censoring

status: 	 censoring status

differ: 	 differentiation of tumour (1=well, 2=moderate, 3=poor)

extent: 	 Extent of local spread (1=submucosa, 2=muscle, 3=serosa, 4=contiguous structures)

surg: 	 time from surgery to registration (0=short, 1=long)

node4: 	 more than 4 positive lymph nodes

etype: 	 event type: 1=recurrence,2=death

Source

Danyu Lin

[Package survival version 2.35-4 Index]