coxph.control {survival}	R Documentation

Ancillary arguments for controling coxph fits

Description

This is used to set various numeric parameters controling a Cox model fit. Typically it would only be used in a call to coxph.

Usage

coxph.control(eps = 1e-08, toler.chol = .Machine$double.eps^0.75, iter.max = 15, toler.inf = sqrt(eps), outer.max = 10)

Arguments

eps 	 Iteration continues until the relative change in the log partial likelihood is less than eps. Must be positive.

toler.chol 	 Tolerance for detection of singularity during a Cholesky decomposion of the variance matrix, i.e., for detecting a redundant predictor variable.

iter.max 	 Maximum number of iterations to attempt for convergence.

toler.inf 	 Tolerance criteria for the warning message about a possible infinite coefficient value.

outer.max 	 For a penalized coxph model, e.g. with pspline terms, there is an outer loop of iteration to determine the penalty parameters; maximum number of iterations for this outer loop.

Value

a list containing the values of each of the above constants

Author(s)

Terry Therneau

See Also

coxph

[Package survival version 2.35-4 Index]