coxph.object {survival}	R Documentation

Proportional Hazards Regression Object

Description

This class of objects is returned by the coxph class of functions to represent a fitted proportional hazards model.

Objects of this class have methods for the functions print, summary, residuals, predict and survfit.

COMPONENTS

The following components must be included in a legitimate coxph object.

coefficients

the coefficients of the linear predictor, which multiply the columns of the model matrix. If the model is over-determined there will be missing values in the vector corresponding to the redundant columns in the model matrix.

var

the variance matrix of the coefficients. Rows and columns corresponding to any missing coefficients are set to zero.

naive.var

this component will be present only if the robust option was true. If so, the var component will contain the robust estimate of variance, and this component will contain the ordinary estimate.

loglik

a vector of length 2 containing the log-likelihood with the initial values and with the final values of the coefficients.

score

value of the efficient score test, at the initial value of the coefficients.

rscore

the robust log-rank statistic, if a robust variance was requested.

wald.test

the Wald test of whether the final coefficients differ from the initial values.

iter

number of iterations used.

linear.predictors

the vector of linear predictors, one per subject.

residuals

the martingale residuals.

means

vector of column means of the X matrix. Subsequent survival curves are adjusted to this value.

n

the number of observations used in the fit.

weights

the vector of case weights, if one was used.

method

the computation method used.

na.action

the na.action attribute, if any, that was returned by the na.action routine.

The object will also contain the following, for documentation see the lm object: terms, assign, formula, call, and, optionally, x, y, and/or frame.

See Also

coxph, coxph.detail, cox.zph, survfit, residuals.coxph, survreg

[Package survival version 2.34-1 Index]