coxph.object {survival}	R Documentation

Proportional Hazards Regression Object

Description

This class of objects is returned by the coxph class of functions to represent a fitted proportional hazards model. Objects of this class have methods for the functions print, summary, residuals, predict and survfit.

Arguments

coefficients
the vector of coefficients. If the model is over-determined there will be missing values in the vector corresponding to the redundant columns in the model matrix.

var
the variance matrix of the coefficients. Rows and columns corresponding to any missing coefficients are set to zero.

naive.var
this component will be present only if the robust option was true. If so, the var component will contain the robust estimate of variance, and this component will contain the ordinary estimate.

loglik
a vector of length 2 containing the log-likelihood with the initial values and with the final values of the coefficients.

score
value of the efficient score test, at the initial value of the coefficients.

rscore
the robust log-rank statistic, if a robust variance was requested.

wald.test
the Wald test of whether the final coefficients differ from the initial values.

iter
number of iterations used.

linear.predictors
the vector of linear predictors, one per subject. Note that this vector has been centered, see predict.coxph for more details.

residuals
the martingale residuals.

means
vector of values used as the reference for each covariate. For instance, a later call to predict(fit, type='risk') will give the hazard ratio between an observation and this reference. (For most covariates this will contain the mean.)

n
the number of observations used in the fit.

nevent
the number of events (usually deaths) used in the fit.

concordance
a vector of length 6, containing the number of pairs that are concordant, discordant, tied on x, tied on y, and tied on both, followed by the standard error of the concordance statistic.

first
the first derivative vector at the solution.

weights
the vector of case weights, if one was used.

method
the method used for handling tied survival times.

na.action
the na.action attribute, if any, that was returned by the na.action routine.

timefix
the value of the timefix option used in the fit

cmap
the coefficient map, present for multi-state coxph fits. There a column for each transition and a row for each coefficient, the value maps that transition/coefficient pair to a position in the coefficient vector. If a particular covariate is not used by the transition the matrix will contain a zero, if two transitions share a coefficient the matrix will contain repeats.

stratum_map
stratum mapping, present for multi-state coxph fits. The row labeled ‘(Baseline)’ identifies transitions that do or do not share a baseline stratum. Further rows correspond to strata() terms in the model, each of which may apply to some transitions and not others.

...
The object will also contain the following, for documentation see the lm object: terms, assign, formula, call, and, optionally, x, y, and/or frame.

Components

The following components must be included in a legitimate coxph object.

See Also

coxph, coxph.detail, cox.zph, residuals.coxph, survfit, survreg.

[Package survival version 3.3-1 Index]