coxph.object {survival}	R Documentation

Proportional Hazards Regression Object

Description

This class of objects is returned by the coxph class of functions to represent a fitted proportional hazards model. Objects of this class have methods for the functions print, summary, residuals, predict and survfit.

Arguments

coefficients
the vector of coefficients. If the model is over-determined there will be missing values in the vector corresponding to the redundant columns in the model matrix.

var
the variance matrix of the coefficients. Rows and columns corresponding to any missing coefficients are set to zero.

naive.var
this component will be present only if the robust option was true. If so, the var component will contain the robust estimate of variance, and this component will contain the ordinary estimate.

loglik
a vector of length 2 containing the log-likelihood with the initial values and with the final values of the coefficients.

score
value of the efficient score test, at the initial value of the coefficients.

rscore
the robust log-rank statistic, if a robust variance was requested.

wald.test
the Wald test of whether the final coefficients differ from the initial values.

iter
number of iterations used.

linear.predictors
the vector of linear predictors, one per subject. Note that this vector has been centered, see predict.coxph for more details.

residuals
the martingale residuals.

means
vector of column means of the X matrix. Subsequent survival curves are adjusted to this value.

n
the number of observations used in the fit.

nevent
the number of events (usually deaths) used in the fit.

concordance
the concordance, as computed by survConcordance.

first
the first derivative vector at the solution.

weights
the vector of case weights, if one was used.

method
the computation method used.

na.action
the na.action attribute, if any, that was returned by the na.action routine.

The object will also contain the following, for documentation see the lm object: terms, assign, formula, call, and, optionally, x, y, and/or frame.

Components

The following components must be included in a legitimate coxph object.

See Also

coxph, coxph.detail, cox.zph, residuals.coxph, survfit, survreg.

[Package survival version 2.42-6 Index]