pbc {survival}	R Documentation

Mayo Clinic Primary Biliary Cirrhosis Data

Description

Followup of 312 randomised and 108 unrandomised patients with primary biliary cirrhosis, a rare autoimmune liver disease, at Mayo Clinic. Most variables have some missing data, coded as -9.

Usage

pbc

Format

age: 	 in years

alb: 	 serum albumin

alkphos: 	 alkaline phosphotase

ascites: 	 presence of ascites

bili: 	 serum bilirunbin

chol: 	 serum cholesterol

edema: 	 presence of edema

edtrt: 	 0 no edema, 0.5 untreated or successfully treated

1 unsuccessfully treated edema

hepmeg: 	 enlarged liver

time: 	 survival time

platelet: 	 platelet count

protime: 	 standardised blood clotting time

sex: 	 1=male

sgot: 	 liver enzyme (now called AST)

spiders: 	 blood vessel malformations in the skin

stage: 	 histologic stage of disease (needs biopsy)

status: 	 censoring status

trt: 	 1/2/-9 for control, treatment, not randomised

trig: 	 triglycerides

copper: 	 urine copper

Source

TR Fleming and DP Harrington (1991), Counting Processes and Survival Analysis. Wiley, New York. ISBN: 0-471-52218-X.

[Package survival version 2.34-1 Index]