pbcseq {survival}	R Documentation

Mayo Clinic Primary Biliary Cirrhosis, sequential data

Description

This data is a continuation of the PBC data set, and contains the follow-up laboratory data for each study patient. An analysis based on the data can be found in Murtagh, et. al.

The primary PBC data set contains only baseline measurements of the laboratory paramters. This data set contains multiple laboratory results, but only on the 312 randomized patients. Some baseline data values in this file differ from the original PBC file, for instance, the data errors in prothrombin time and age which were discovered after the orignal analysis (see Fleming and Harrington, figure 4.6.7).

One "feature" of the data deserves special comment. The last observation before death or liver transplant often has many more missing covariates than other data rows. The original clinical protocol for these patients specified visits at 6 months, 1 year, and annually thereafter. At these protocol visits lab values were obtained for a large pre-specified battery of tests. "Extra" visits, often undertaken because of worsening medical condition, did not necessarily have all this lab work. The missing values are thus potentially informative.

Usage

pbc

Format

id: 	 case number

age: 	 in years

sex: 	 m/f

trt: 	 1/2/NA for D-penicillmain, placebo, not randomised

time: 	 number of days between registration and the earlier of death,

transplantion, or study analysis in July, 1986

status: 	 status at endpoint, 0/1/2 for censored, transplant, dead

day: 	 number of days between enrollment and this visit date

all measurements below refer to this date

albumin: 	 serum albumin (mg/dl)

alk.phos: 	 alkaline phosphotase (U/liter)

ascites: 	 presence of ascites

ast: 	 aspartate aminotransferase, once called SGOT (U/ml)

bili: 	 serum bilirunbin (mg/dl)

chol: 	 serum cholesterol (mg/dl)

copper: 	 urine copper (ug/day)

edema: 	 0 no edema, 0.5 untreated or successfully treated

1 edema despite diuretic therapy

hepato: 	 presence of hepatomegaly or enlarged liver

platelet: 	 platelet count

protime: 	 standardised blood clotting time

spiders: 	 blood vessel malformations in the skin

stage: 	 histologic stage of disease (needs biopsy)

trig: 	 triglycerides (mg/dl)

Source

T Therneau and P Grambsch, "Modeling Survival Data: Extending the Cox Model", Springer-Verlag, New York, 2000. ISBN: 0-387-98784-3.

References

Murtaugh PA. Dickson ER. Van Dam GM. Malinchoc M. Grambsch PM. Langworthy AL. Gips CH. "Primary biliary cirrhosis: prediction of short-term survival based on repeated patient visits." Hepatology. 20(1.1):126-34, 1994.

Fleming T and Harrington D., "Counting Processes and Survival Analysis", Wiley, New York, 1991.

[Package survival version 2.35-4 Index]