plot.aareg {survival}	R Documentation

Plot an aareg object.

Description

Plot the estimated coefficient function(s) from a fit of Aalen's additive regression model.

Usage

## S3 method for class 'aareg':
plot(x, se=TRUE, maxtime, type='s', ...)

Arguments

x 	 the result of a call to the aareg function

se 	 if TRUE, standard error bands are included on the plot

maxtime 	 upper limit for the x-axis.

type 	 graphical parameter for the type of line, default is "steps".

... 	 other graphical parameters such as line type, color, or axis labels.

Side Effects

A plot is produced on the current graphical device.

References

Aalen, O.O. (1989). A linear regression model for the analysis of life times. Statistics in Medicine, 8:907-925.

See Also

aareg

[Package survival version 2.35-4 Index]