print.summary.coxph {survival}	R Documentation

Print method for summary.coxph objects

Description

Produces a printed summary of a fitted coxph model

Usage

## S3 method for class 'summary.coxph'
print(x, digits=max(getOption("digits") - 3, 3),
  signif.stars = getOption("show.signif.stars"), expand=FALSE, ...)

Arguments

x
the result of a call to summary.coxph

digits
significant digits to print

signif.stars
Show stars to highlight small p-values

expand
if the summary is for a multi-state coxph fit, print the results in an expanded format.

...
For future methods

[Package survival version 3.3-1 Index]