print.summary.survexp {survival}	R Documentation

Print Survexp Summary

Description

Prints the results of summary.survexp

Usage


## S3 method for class 'summary.survexp'
print(x, digits = max(options()$digits - 4, 3), ...)

Arguments

x
an object of class summary.survexp.

digits
the number of digits to use in printing the result.

...
for future methods

Value

x, with the invisible flag set to prevent further printing.

Author(s)

Terry Therneau

See Also

link{summary.survexp}, survexp

[Package survival version 2.42-6 Index]