print.summary.survfit {survival}	R Documentation

Print Survfit Summary

Description

Prints the result of summary.survfit.

Usage


## S3 method for class 'summary.survfit'
print(x, digits = max(options() $digits-4, 3), ...)

Arguments

x
an object of class "summary.survfit", which is the result of the summary.survfit function.

digits
the number of digits to use in printing the numbers.

...
for future methods

Value

x, with the invisible flag set to prevent printing.

Side Effects

prints the summary created by summary.survfit.

See Also

options, print, summary.survfit.

[Package survival version 2.42-6 Index]