pspline {survival}	R Documentation

Penalised smoothing splines

Description

Specifies a penalised spline basis for the predictor. This is done by fitting a comparatively small set of splines and penalising the integrated second derivative. Results are similar to smoothing splines with a knot at each data point but computationally simpler.

Usage

pspline(x, df=4, theta, nterm=2.5 * df, degree=3, eps=0.1, method, ...)

Arguments

x 	 predictor

df 	 approximate degrees of freedom. df=0 means use AIC

theta 	 roughness penalty

nterm 	 number of splines in the basis

degree 	 degree of splines

eps 	 accuracy for df

method 	 Method for automatic choice of theta

... 	 I don't know what this does

Value

Object of class coxph.penalty containing the spline basis with attributes specifying control functions.

See Also

coxph,survreg,ridge,frailty

Examples

lfit6 <- survreg(Surv(time, status)~pspline(age, df=2), cancer)
plot(cancer$age, predict(lfit6), xlab='Age', ylab="Spline prediction")
title("Cancer Data")
fit0 <- coxph(Surv(time, status) ~ ph.ecog + age, cancer)
fit1 <- coxph(Surv(time, status) ~ ph.ecog + pspline(age,3), cancer)
fit3 <- coxph(Surv(time, status) ~ ph.ecog + pspline(age,8), cancer)
fit0
fit1
fit3

[Package survival version 2.34-1 Index]