ratetable {survival}	R Documentation

Allow ratetable() terms in a model

Description

This function supports ratetable() terms in a model statement, within survexp and pyears.

Usage

ratetable(...)

Arguments

...
the named dimensions of a rate table

Details

This way of mapping a rate table's variable names to a user data frame has been superseded, instead use the rmap argument of the survexp, pyears, or survdiff routines. The function remains only to allow older code to be run.

Author(s)

Terry Therneau

[Package survival version 3.5-3 Index]