ratetable {survival}	R Documentation

Ratetable reference in formula

Description

This function matches variable names in data to those in a ratetable for survexp

Usage

ratetable(...)

Arguments

...
tags matching dimensions of the ratetable and variables in the data frame (see example)

Value

A data frame

See Also

survexp,survexp.us,is.ratetable

Examples

fit <- survfit(Surv(time, status) ~ sex, pbc,subset=1:312)

# The data set does not have entry date, use the midpoint of the study
efit <- survexp(~ ratetable(sex=sex,age=age*365.35,year=as.Date('1979/1/1')) +
            sex, data=pbc,  times=(0:24)*182)
## Not run: 
plot(fit, mark.time=F, xscale=365.25, xlab="Years post diagnosis",
        ylab="Survival")
lines(efit, col=2) # Add the expected survival line

## End(Not run)

[Package survival version 2.41-3 Index]