Ridge regression

When used in a coxph or survreg model formula, specifies a ridge regression term. The likelihood is penalised by theta/2 time the sum of squared coefficients. If scale=T the penalty is calculated for coefficients based on rescaling the predictors to have unit variance. If df is specified then theta is chosen based on an approximate degrees of freedom.

Usage

ridge(..., theta, df=nvar/2, eps=0.1, scale=TRUE)

Arguments

...

predictors to be ridged

theta

penalty is theta/2 time sum of squared coefficients

df

Approximate degrees of freedom

eps

Accuracy required for df

scale

Scale variables before applying penalty?

Note

If the expression ridge(x1, x2, x3, ...) is too many characters long then the internal terms() function will add newlines to the variable name and then the coxph routine simply gets lost. (Some labels will have the newline and some won't.) One solution is to bundle all of the variables into a single matrix and use that matrix as the argument to ridge so as to shorten the call, e.g. mdata$many <- as.matrix(mydata[,5:53]).

Value

An object of class coxph.penalty containing the data and control functions.

References

Gray (1992) "Flexible methods of analysing survival data using splines, with applications to breast cancer prognosis" JASA 87:942--951

See also

Examples

coxph(Surv(futime, fustat) ~ rx + ridge(age, ecog.ps, theta=1),
        ovarian)
#> Call:
#> coxph(formula = Surv(futime, fustat) ~ rx + ridge(age, ecog.ps, 
#>     theta = 1), data = ovarian)
#> 
#>                   coef se(coef)     se2   Chisq DF      p
#> rx             -0.8564   0.6161  0.6156  1.9323  1 0.1645
#> ridge(age)      0.1229   0.0385  0.0354 10.2127  1 0.0014
#> ridge(ecog.ps)  0.1093   0.5734  0.5484  0.0363  1 0.8488
#> 
#> Iterations: 1 outer, 5 Newton-Raphson
#> Degrees of freedom for terms= 1.0 1.8 
#> Likelihood ratio test=15.6  on 2.76 df, p=0.001
#> n= 26, number of events= 12 

lfit0 <- survreg(Surv(time, status) ~1, lung)
lfit1 <- survreg(Surv(time, status) ~ age + ridge(ph.ecog, theta=5), lung)
lfit2 <- survreg(Surv(time, status) ~ sex + ridge(age, ph.ecog, theta=1), lung)
lfit3 <- survreg(Surv(time, status) ~ sex + age + ph.ecog, lung)