strata {survival}	R Documentation

Identify Stratification Variables

Description

This is a special function used in the context of the Cox survival model. It identifies stratification variables when they appear on the right hand side of a formula.

Usage

strata(..., na.group=FALSE, shortlabel=FALSE, sep=', ')

Arguments

... 	 any number of variables. All must be the same length.

na.group 	 a logical variable, if TRUE, then missing values are treated as a distinct level of each variable.

shortlabel 	 if TRUE omit variable names from resulting factor labels

sep 	 the character used to separate groups, in the created label

Details

The result is identical to the interaction function, but for the labeling of the factors (strata is more verbose).

Value

a new factor, whose levels are all possible combinations of the factors supplied as arguments.

See Also

coxph, interaction

Examples

a<-factor(rep(1:3,4))
b<-factor(rep(1:4,3))
levels(strata(a))
levels(strata(a,b,shortlabel=TRUE))

coxph(Surv(futime, fustat) ~ age + strata(rx), data=ovarian) 

[Package survival version 2.35-4 Index]