Summary method for Cox models

Produces a summary of a fitted coxph model

Usage

# S3 method for coxph
summary(object, conf.int=0.95, scale=1,...)

Arguments

object

the result of a coxph fit

conf.int

level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed

scale

vector of scale factors for the coefficients, defaults to 1. The printed coefficients, se, and confidence intervals will be associated with one scale unit.

...

for future methods

Value

An object of class summary.coxph, with components:

n, nevent

number of observations and number of events, respectively, in the fit

loglik

the log partial likelihood at the initial and final values

coefficients

a matrix with one row for each coefficient, and columns containing the coefficient, the hazard ratio exp(coef), standard error, Wald statistic, and P value.

conf.int

a matrix with one row for each coefficient, containing the confidence limits for exp(coef)

logtest, sctest, waldtest

the overall likelihood ratio, score, and Wald test statistics for the model

concordance

the concordance statistic and its standard error

used.robust

whether an asymptotic or robust variance was used

rsq

an approximate R^2 based on Nagelkirke (Biometrika 1991).

fail

a message, if the underlying coxph call failed

call

a copy of the call

na.action

information on missing values

Note

The pseudo r-squared of Nagelkirke is attractive because it is simple, but further work has shown that it has poor properties and it is now deprecated. The value is no longer printed by default, and will eventually be removed from the object.

See also

Examples

fit <- coxph(Surv(time, status) ~ age + sex, lung) 
summary(fit)
#> Call:
#> coxph(formula = Surv(time, status) ~ age + sex, data = lung)
#> 
#>   n= 228, number of events= 165 
#> 
#>          coef exp(coef)  se(coef)      z Pr(>|z|)   
#> age  0.017045  1.017191  0.009223  1.848  0.06459 . 
#> sex -0.513219  0.598566  0.167458 -3.065  0.00218 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#>     exp(coef) exp(-coef) lower .95 upper .95
#> age    1.0172     0.9831    0.9990    1.0357
#> sex    0.5986     1.6707    0.4311    0.8311
#> 
#> Concordance= 0.603  (se = 0.025 )
#> Likelihood ratio test= 14.12  on 2 df,   p=9e-04
#> Wald test            = 13.47  on 2 df,   p=0.001
#> Score (logrank) test = 13.72  on 2 df,   p=0.001
#>