summary.coxph {survival}	R Documentation

Summary method for Cox models

Description

Produces a summary of a fitted coxph model

Usage

## S3 method for class 'coxph'
summary(object, conf.int=0.95, scale=1,...)

Arguments

object
the result of a coxph fit

conf.int
level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed

scale
vector of scale factors for the coefficients, defaults to 1. The printed coefficients, se, and confidence intervals will be associated with one scale unit.

...
for future methods

Value

An object of class summary.coxph, with components:

n, nevent
number of observations and number of events, respectively, in the fit

loglik
the log partial likelihood at the initial and final values

coefficients
a matrix with one row for each coefficient, and columns containing the coefficient, the hazard ratio exp(coef), standard error, Wald statistic, and P value.

conf.int
a matrix with one row for each coefficient, containing the confidence limits for exp(coef)

logtest, sctest, waldtest
the overall likelihood ratio, score, and Wald test statistics for the model

concordance
the concordance statistic and its standard error

used.robust
whether an asymptotic or robust variance was used

rsq
an approximate R^2 based on Nagelkirke (Biometrika 1991).

fail
a message, if the underlying coxph call failed

call
a copy of the call

na.action
information on missing values

Note

The pseudo r-squared of Nagelkirke is attractive because it is simple, but further work has shown that it has poor properties and it is now depricated. The value is no longer printed by default, and will eventually be removed from the object.

See Also

coxph, print.coxph

Examples

fit <- coxph(Surv(time, status) ~ age + sex, lung) 
summary(fit)

[Package survival version 3.1-12 Index]