summary.survfit {survival}	R Documentation

Summary of a Survival Curve

Description

Returns a list containing the survival curve, confidence limits for the curve, and other information.

Usage

## S3 method for class 'survfit':
summary(object, times, censored=FALSE, scale=1, ...)

Arguments

object 	 output from a call to survfit.

times 	 vector of times; the returned matrix will contain 1 row for each time. This must be in increasing order and missing values are not allowed. If censored=T, the default times vector contains all the unique times in fit, otherwise the default times vector uses only the event (death) times.

censored 	 logical flag: should the censoring times be included in the output? This is ignored if the times argument is present.

scale 	 rescale the survival time, e.g., if the input data to survfit were in days, scale=365.25 would scale the output to years.

... 	 other unused arguments

Value

a list with the following components

time 	 the timepoint on the curve.

surv 	 the value of the survival curve at time t+0.

n.risk 	 the number of subjects at risk at time t-0 (but see the comments on weights in the survfit help file).

n.event 	 if the times argument is missing, then this column is the number of events that occurred at time t. Otherwise, it is the cumulative number of events that have occurred since the last time listed until time t+0.

std.err 	 the standard error of the survival value.

conf.int 	 level of confidence for the confidence intervals of survival.

lower 	 lower confidence limits for the curve.

upper 	 upper confidence limits for the curve.

strata 	 indicates stratification of curve estimation. If strata is not NULL, there are multiple curves in the result and the surv, time, n.risk, etc. vectors will contain multiple curves, pasted end to end. The levels of strata (a factor) are the labels for the curves.

call 	 the statement used to create the fit object.

na.action 	 passed through from fit, if present.

See Also

survfit, print.summary.survfit.

Examples

summary( survfit( Surv(futime, fustat),data=ovarian))
summary( survfit( Surv(futime, fustat)~rx,data=ovarian))

[Package survival version 2.34-1 Index]