survfit.object {survival}	R Documentation

Survival Curve Object

Description

This class of objects is returned by the survfit class of functions to represent a fitted survival curve. For a multi-state model the object has class c('survfitms', 'survfit').

Objects of this class have methods for the functions print, summary, plot, points and lines. The print.survfit method does more computation than is typical for a print method and is documented on a separate page.

Arguments

n
total number of subjects in each curve.

time
the time points at which the curve has a step.

n.risk
the number of subjects at risk at t.

n.event
the number of events that occur at time t.

n.enter
for counting process data only, the number of subjects that enter at time t.

n.censor
for counting process data only, the number of subjects who exit the risk set, without an event, at time t. (For right censored data, this number can be computed from the successive values of the number at risk).

surv
the estimate of survival at time t+0. This may be a vector or a matrix. The latter occurs when a set of survival curves is created from a single Cox model, in which case there is one column for each covariate set.

pstate
a multi-state survival will have the pstate component instead of surv. It will be a matrix containing the estimated probability of each state at each time, one column per state.

std.err
for a survival curve this contains standard error of the cumulative hazard or -log(survival), for a multi-state curve it contains the standard error of prev. This difference is a reflection of the fact that each is the natural calculation for that case.

cumhaz hazard
optional. Contains the cumulative hazard for each possible transtion.

strata
if there are multiple curves, this component gives the number of elements of the time vector corresponding to the first curve, the second curve, and so on. The names of the elements are labels for the curves.

upper
optional upper confidence limit for the survival curve or pstate

lower
options lower confidence limit for the survival curve or pstate

start.time, p0, sp0
see details below

newdata
for survival curves from a fitted model, this contains the covariate values for the curves

n.all
the total number of observations that were available For counting process data, and any time that the start.time argument was used, not all may have been used in creating the curve, in which case this value will be larger than n above. The print and plot routines in the package do no use this value, it is for information only.

conf.type
the approximation used to compute the confidence limits.

conf.int
the level of the confidence limits, e.g. 90 or 95%.

transitions
for multi-state data, the total number of transitions of each type.

na.action
the returned value from the na.action function, if any. It will be used in the printout of the curve, e.g., the number of observations deleted due to missing values.

call
an image of the call that produced the object.

type
type of survival censoring.

influence.p, influence.c
optional influence matrices for the pstate (or surv) and for the cumhaz estimates. A list with one element per stratum, each element of the list is an array indexed by subject, time, state.

version
the version of the object. Will be missing, 2, or 3

Structure

The following components must be included in a legitimate survfit or survfitms object.

Subscripts

Survfit objects can be subscripted. This is often used to plot a subset of the curves, for instance. From the user's point of view the survfit object appears to be a vector, matrix, or array of curves. The first dimension is always the underlying number of curves or “strata”; for multi-state models the state is always the last dimension. Predicted curves from a Cox model can have a second dimension which is the number of different covariate prediction vectors.

Details

The survfit object has evolved over time: when first created there was no thought of multi-state models for instance. This evolution has almost entirely been accomplished by the addition of new elements. One change in survival version 3 is the addition of a survfitconf routine which will compute confidence intervals for a survfit object. This allows the computation of CI intervals to be deferred until later, if desired, rather than making them a permanent part of the object. Later iterations of the base routines may omit the confidence intervals.

The survfit object starts at the first observation time, but survival curves are normally plotted from time 0. A helper routine survfit0 can be used to add this first time point and align the data.

See Also

plot.survfit, summary.survfit, print.survfit, survfit, survfit0

[Package survival version 3.1-12 Index]