survival-internal {survival}	R Documentation

Internal survival functions

Description

Internal survival functions

Usage

survreg.fit(x, y, weights, offset, init, controlvals, dist, scale = 0,
    nstrat = 1, strata, parms = NULL)
survpenal.fit(x, y, weights, offset, init, controlvals, dist, scale = 0,
    nstrat = 1, strata, pcols, pattr, assign, parms = NULL)
survdiff.fit(y, x, strat, rho = 0)
is.category(x)
match.ratetable(R, ratetable)
## S3 method for class 'ratetable':
as.matrix(x, ...)
is.na.ratetable2(x)
is.na.coxph.penalty(x)
model.newframe(object, newdata, safe = FALSE, response = FALSE, ...)
## S3 method for class 'difftime':
as.data.frame(x, row.names = NULL, optional = FALSE,...)
coxpenal.df(hmat, hinv, fdiag, assign.list, ptype, nvar, pen1,
    pen2, sparse)
coxpenal.fit(x, y, strata, offset, init, control, weights, method,
    rownames, pcols, pattr, assign)
coxph.fit(x, y, strata, offset, init, control, weights, method,
    rownames)
coxph.wtest(var, b, toler.chol = 1e-09)
agreg.fit(x, y, strata, offset, init, control, weights, method,
    rownames)
agexact.fit(x, y, strata, offset, init, control, weights, method,
    rownames) 

Details

These are not to be called by the user

[Package survival version 2.34-1 Index]