survival-internal {survival}	R Documentation

Internal survival functions

Description

Internal survival functions

Usage

survreg.fit(x, y, weights, offset, init, controlvals, dist, scale = 0,
    nstrat = 1, strata, parms = NULL,assign)
survpenal.fit(x, y, weights, offset, init, controlvals, dist, scale = 0,
    nstrat = 1, strata, pcols, pattr, assign, parms = NULL)
survdiff.fit(y, x, strat, rho = 0)
match.ratetable(R, ratetable)
## S3 method for class 'ratetable'
as.matrix(x, ...)
## S3 method for class 'coxph.penalty'
is.na(x)
coxpenal.df(hmat, hinv, fdiag, assign.list, ptype, nvar, pen1,
    pen2, sparse)
coxpenal.fit(x, y, strata, offset, init, control, weights, method,
    rownames, pcols, pattr, assign, nocenter)
coxph.wtest(var, b, toler.chol = 1e-09)
agexact.fit(x, y, strata, offset, init, control, weights, method,
    rownames, resid=TRUE, nocenter=NULL) 
survfitCI(X, Y, weights,  id, cluster, robust, istate,
                       stype=1, ctype=1, 
                       se.fit=TRUE,
                       conf.int= .95,
                       conf.type=c('log',  'log-log',  'plain', 'none',
                                   'logit', 'arcsin'),
                       conf.lower=c('usual', 'peto', 'modified'),
                       influence=FALSE, start.time, p0, type)
survfitKM(x, y, weights=rep(1,length(x)),
                       stype=1, ctype=1, 
                       se.fit=TRUE,
                       conf.int= .95,
                       conf.type=c('log',  'log-log',  'plain', 'none',
                                   'logit', 'arcsin'),
                       conf.lower=c('usual', 'peto', 'modified'),
                       start.time, id, cluster, robust,
                       influence=FALSE, type) 
survfitTurnbull(x, y, weights,
                       type=c('kaplan-meier', 'fleming-harrington', 'fh2'),
                       error=c('greenwood', "tsiatis"), se.fit=TRUE,
                       conf.int= .95,
                       conf.type=c('log',  'log-log',  'plain', 'none',
                                   'logit', 'arcsin'),
                       conf.lower=c('usual', 'peto', 'modified'),
                       start.time, robust, cluster)

Details

The arguments to these routines are not guaranteed to stay the same from release to release – call them at your own risk!

[Package survival version 3.3-1 Index]