survreg {survival}	R Documentation

Regression for a Parametric Survival Model

Description

Regression for a parametric survival model. These are all time-transformed location models, with the most useful case being the accelerated failure models that use a log transformation.

Usage

survreg(formula=formula(data), data=parent.frame(), weights, 
subset,na.action,dist="weibull",  init=NULL, scale=0,
 control=survreg.control(),parms=NULL,model=FALSE, x=FALSE,
 y=TRUE, robust=FALSE, ...)

Arguments

formula 	 a formula expression as for other regression models. See the documentation for lm and formula for details.

data 	 optional data frame in which to interpret the variables occurring in the formula.

weights 	 Optional observation weights

subset 	 subset of the observations to be used in the fit.

na.action 	 function to be used to handle any NAs in the data.

dist 	 assumed distribution for y variable.
If the argument is a character string, then it is assumed to name an element from survreg.distributions. These include "weibull", "exponential", "gaussian", "logistic", "lognormal" and "loglogistic". Only enough of the name needs to be given to make the choice unique.
Otherwise, it is assumed to be a user defined list conforming to the format described in survreg.distributions.

parms 	 a list of fixed parameters. For the t-distribution for instance this is the degrees of freedom; most of the distributions have no parameters.

init 	 optional vector of initial values for the parameters.

scale 	 optional fixed value for the scale. If set to <=0 then the scale is estimated.

control 	 a list of control values, in the format producted by survreg.control.

model 	 if TRUE, the model frame is returned.

x 	 if TRUE, then the X matrix is returned.

y 	 if TRUE, then the y vector (or survival times) is returned.

robust 	 if TRUE, sandwich standard errors are computed. Defaults to TRUE when formula contains a cluster term.

... 	 other arguments which will be passed to survreg.control.

Value

an object of class survreg is returned.

Compatibility note

This routine underwent significant changes from survival4 to survival5. The survreg.old function gives a backwards-compatible interface. In S-PLUS the new function is called survReg and the old one survreg.

See Also

survreg.object, survreg.distributions, pspline, frailty, ridge, survreg.old

Examples

## These are all the same
survreg(Surv(futime, fustat) ~ ecog.ps + rx, ovarian, dist='weibull',scale=1)
survreg(Surv(futime, fustat) ~ ecog.ps + rx, ovarian,
dist="exponential")
survreg.old(Surv(futime, fustat) ~ ecog.ps + rx, ovarian, dist='extreme',fixed=list(scale=1),link="log")

[Package survival version 2.34-1 Index]