survreg.control {survival}	R Documentation

Package options for survreg and coxph

Description

This functions checks and packages the fitting options for survreg

Usage

survreg.control(maxiter=30, rel.tolerance=1e-09, 
toler.chol=1e-10, iter.max, debug=0, outer.max=10)

Arguments

maxiter 	 maximum number of iterations

rel.tolerance 	 relative tolerance to declare convergence

toler.chol 	 Tolerance to declare Cholesky decomposition singular

iter.max 	 same as maxiter

debug 	 print debugging information

outer.max 	 maximum number of outer iterations for choosing penalty parameters

Value

A list with the same elements as the input

See Also

survreg

[Package survival version 2.35-4 Index]