survreg.object {survival}	R Documentation

Parametric Survival Model Object

Description

This class of objects is returned by the survreg function to represent a fitted parametric survival model. Objects of this class have methods for the functions print, summary, predict, and residuals.

COMPONENTS

The following components must be included in a legitimate survreg object.

coefficients

the coefficients of the linear.predictors, which multiply the columns of the model matrix. It does not include the estimate of error (sigma). The names of the coefficients are the names of the single-degree-of-freedom effects (the columns of the model matrix). If the model is over-determined there will be missing values in the coefficients corresponding to non-estimable coefficients.

icoef

coefficients of the baseline model, which will contain the intercept and log(scale), or mulitple scale factors for a stratified model.

var

the variance-covariance matrix for the parameters, including the log(scale) parameter(s).

loglik

a vector of length 2, containing the log-likelihood for the baseline and full models.

iter

the number of iterations required

linear.predictors

the linear predictor for each subject.

df

the degrees of freedom for the final model. For a penalized model this will be a vector with one element per term.

scale

the scale factor(s), with length equal to the number of strata.

idf

degrees of freedom for the initial model.

means

a vector of the column means of the coefficient matrix.

dist

the distribution used in the fit.

The object will also have the following components found in other model results (some are optional): linear predictors, weights, x, y, model, call, terms and formula. See lm.
See Also

survreg, lm

[Package survival version 2.35-4 Index]