untangle.specials {survival}	R Documentation

Help Process the ‘specials’ Argument of the ‘terms’ Function.

Description

Given a terms structure and a desired special name, this returns an index appropriate for subscripting the terms structure and another appropriate for the data frame.

Usage

untangle.specials(tt, special, order=1)

Arguments

tt 	 a terms object.

special 	 the name of a special function, presumably used in the terms object.

order 	 the order of the desired terms. If set to 2, interactions with the special function will be included.

Value

a list with two components:

vars 	 a vector of variable names, as would be found in the data frame, of the specials.

terms 	 a numeric vector, suitable for subscripting the terms structure, that indexes the terms in the expanded model formula which involve the special.

Examples

formula<-Surv(tt,ss)~x+z*strata(id)
tms<-terms(formula,specials="strata")
## the specials attribute
attr(tms,"specials")
## main effects 
untangle.specials(tms,"strata")
## and interactions
untangle.specials(tms,"strata",order=1:2)

[Package survival version 2.35-4 Index]